Correct a lot of bugs. It works!

+++ run-experiments.py
@@ -16,12 +16,15 @@ parser.add_argument('configs_dir', type=str, help='Path to configs directory')
 parser.add_argument('output_dir', type=str, help='Path to the output directory')
 parser.add_argument('-v', choices=['0','1','2','3'], default=1, dest='verbose',
                     help='Level of verbosity, between 0 and 3')
+parser.add_argument('-l', '--load', action="store_true", help='Load the last model for already existing experiments')
+
 args = parser.parse_args()
 
 output_dir = os.path.abspath(args.output_dir)
 configs_dir = os.path.abspath(args.configs_dir)
 verbose = args.verbose
-     
+load = args.load
+
 # ================ Load config files =============
 
 config_files = glob.glob(os.path.join(configs_dir, '*.yaml'))
@@ -44,22 +47,22 @@ for i_exp in range(len(config_files)):
     config = load_config(config_files[i_exp])
     exp_output_dir = os.path.join(output_dir, exp_name)
 
-    if os.path.exists(exp_output_dir):
-        print("Output path already exists for this experiment. Do you want to delete it? [y/n]")
-        # answer = str(input())
-        # while answer != 'y' and answer != 'n':
-        #     answer = input()
-        answer = "y"
-        if answer == 'y':
-            shutil.rmtree(exp_output_dir)
-        else:
-            exit()
-    os.makedirs(exp_output_dir)
+    if not os.path.exists(exp_output_dir) or (os.path.exists(exp_output_dir) and not load):
+        if os.path.exists(exp_output_dir) and not load:
+            print("Output path already exists for this experiment. Do you want to delete it? [y/n]")
+            # answer = str(input())
+            # while answer != 'y' and answer != 'n':
+            #     answer = input()
+            answer = "y"
+            if answer == 'y':
+                shutil.rmtree(exp_output_dir)
+        os.makedirs(exp_output_dir)
+        shutil.copy(config_files[i_exp], exp_output_dir)
     
     if verbose >= 1:
         print("- Starting experiment...\n")
         
-    experiment(config, exp_output_dir, verbose=verbose)
+    experiment(config, exp_output_dir, load=load, verbose=verbose)
 
 if verbose >= 1:
     print("\n=================== Done ===================\n")
